backend/requirements.txt
# FastAPI and ASGI server
fastapi[all]>=0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
pgvector==0.2.4
alembic==1.13.1

# Pydantic settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Async tasks
celery[redis]==5.3.6
redis==5.0.1

# Other utilities
httpx==0.26.0

# Vertex AI and Embeddings
google-cloud-aiplatform>=1.38.0
# sentence-transformers installed manually in Dockerfile to enforce CPU only
# sentence-transformers>=2.2.2
# torch and torchvision are installed manually in Dockerfile (CPU version)
# to save build time and space. Do not uncomment.
# torch>=2.0.0
# torchvision>=0.16.0

# SVG Processing (Force rebuild)
svgwrite==1.4.3
Pillow>=10.0.0
numpy>=1.26.0

# Payments
stripe==7.9.0

# Scraping
beautifulsoup4==4.12.3

# AI
google-generativeai>=0.3.2
replicate>=0.25.1
# pypotrace removed: C extension unreliable on Python 3.12.
# Vectorization now uses the 'potrace' system binary (installed via apt in Dockerfile).
